pyproject.toml
[project]
name = "docs"
version = "0.1.0"
description = "Add a short description here"
authors = [
    { name = "herring101", email = "herring101@outlook.jp" }
]
dependencies = [
    "h5py~=3.9.0",
    "numpy~=1.24.4",
    "openai~=0.27.8",
    "python-dotenv~=1.0.0",
    "scikit-learn~=1.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true
